Created login page and moved homepage
Created a login page with all necessary methods and validations.
Moved homepage to pages folder to mantain the project's structure clean.
Modified auth service to store the token of the object in localStorage and to set the client id if the login is successfull.
Created a clientId variable on environment to store it along all the app.
Created a client id setter in auth service to set the client's id on the environment variable.
Minor change on auth controller. It now works as it should.
Updated routing module to contain a route to the login.