Created login guard, and changes to homepage
Created a login guard to prevent user from accesing the homepage if his not logged in.
Created a main page component to contain what homepage contained before. Homepage now contains the navbar and a router outlet.
Updated routerlinks in navbar for it to work properly.
Updated routing module to let homepage use the login guard with canActivate and to set it's children routes.